Integrate project with CI (#1)
- Travis-CI
- Codeclimate
- Hakiri

To enable the project to build successfully on Travis-CI we also had to allow the `Gemfile.lock` to be added to project

The error we were getting was

```bash
Bundler could not find compatible versions for gem "bundler":
  In Gemfile:
    defra_ruby_email was resolved to 0.0.1, which depends on
      rails (~> 4.2.11.1) was resolved to 4.2.11.1, which depends on
        bundler (< 2.0, >= 1.3.0)
  Current Bundler version:
    bundler (2.1.4)
This Gemfile requires a different version of Bundler.
Perhaps you need to update Bundler by running `gem install bundler`?
Could not find gem 'bundler (< 2.0, >= 1.3.0)', which is required by gem 'rails
(~> 4.2.11.1)', in any of the sources.
The command "eval bundle install --jobs=3 --retry=3 --path=${BUNDLE_PATH:-vendor/bundle} " failed. Retrying, 2 of 3.
```

Allowing the Gemfile.lock to be added resolved the issue.