Add ability to enable test route through config (#2)
When we come to build the actual last email route, we want to be able to enable and disable it. For example, this is something we want to be able to use in non-production environments, but must be inaccessible in production.

Our standard pattern is to use configuration to manage this, which would be set by the host application the engine is mounted in.

This change adds the initial config, and an example of the enable/disable feature.