Add last email observer and cache (#3)
The functionality of capturing and displaying the last email is based on work originally done to the [waste-exemptions-service](https://github.com/DEFRA/waste-exemptions-engine/pull/112).

In that, a singleton service was created that would cache the last email sent. It depends on a function we found when researching this problem; `ActionMailer::Base.register_interceptor()`.

[Intercepting emails](https://guides.rubyonrails.org/action_mailer_basics.html#intercepting-emails)

> Interceptors allow you to make modifications to emails before they are handed off to the delivery agents. An interceptor class must implement the `:delivering_email(message)` method which will be called before the email is sent.

So Rails passes each email to be sent to our interceptor, where we then cache the message using a singleton class.

If you read further though, you can see Rails also offers the ability to [observe emails](https://guides.rubyonrails.org/action_mailer_basics.html#observing-emails)

> Observers give you access to the email message after it has been sent. An observer class must implement the `:delivered_email(message)` method, which will be called after the email is sent.

As we have no intention to change the email before it is sent, this change will add a last email observer and cache using the WEX version as a base.