Remove unused elements
Being a rails engine when generated it assumes you may be providing additional pages to an existing application.

However we are just adding a single endpoint that returns JSON, so there are a number of things we simply don't need in the project.

This changes removes those elements so its clearer what is actually needed and used by the project.