Get last email body in all scenarios
In PR #3 where we added the `LastEmailCache` class, we included logic to handle multi-part and standard emails.

To be honest, part of this was due to a memory of having to code for something similar in one of our other projects. We have since tracked the [change we made to WCR frontend](https://github.com/DEFRA/waste-carriers-frontend/pull/228) down. The possible scenario we need to cater for is even more complex.

So this change is about taking the lessons learned in that PR and applying them to our engine.