Refactor tests and mailer to cater for all types
Based on the additional information in found in the WCR version of this code and its tests, to be sure we can capture the details of all email formats we need to expand things.

So now we have a clear understanding of multi-part versus basic, and plain text vs html. With ths knowledge we ensure both are covered in the spec plus we havve refactored `TestMailer` to produce all the kinds of emails we need.

One thing to note is that the WCR last-email spec and code said what we have done here would not work

> As far as I could see we should be able to generate both standard and multipart emails dynamically. Below is an example I found in a couple of places
>
```ruby
mail(to: 'example@example.com', subject: 'Welcome to My Awesome Site') do |format|
  format.html { render html: '<h1>Welcome XYZ!</h1>'.html_safe }
  format.text { render plain: 'Welcome XYZ' }
end
```
>
> So this shouldn't need to rely on a template existing. However no matter what changes I made I always got ActionView::MissingTemplate. And it seems I'm not the only one to get this result.

However we have been able to create multi-part emails in both formats with no need to rely on any views. This is just something to note in case it turns out we actually can't!