Get last email body in all scenarios (#6)
In PR #3 where we added the `LastEmailCache` class, we included logic to handle multi-part and standard emails.

To be honest, part of this was due to a memory of having to code for something similar in one of our other projects. We have since tracked the [change we made to WCR frontend](https://github.com/DEFRA/waste-carriers-frontend/pull/228) down. The possible scenarios we need to cater to is even more complex.

Following taken from WCR frontend PR 228

> It turns out my understanding of how the main content or 'body' of an email is stored in the email message was flawed. It's actually even more complex as you have to factor in whether an attachment has been added or not.
>
> If you've created a multipart email then you'll have both a text and a html version (determined by adding the relevant erb views). If you do so then my_mail.parts.length will at least equal 2. If you set only one of them when creating the email then parts doesn't get populated. ActionMailer just populates the root body and sets the content type accordingly.
>
> However any attachments will cause ActionMailer to use parts irrespective of whether you set one, both or neither! So for example if we have a text only email with an attached image, then parts will be of length 2; one being the content and the other being the attachment.