Add rake task to test email connections
Dotted around some of our services a rake tasks that allow us to test if email is working. These have proved handy when updating or moving to a new environment as a way of checking that emails can be sent successfully from the service.

Now we have this gem, and we needed to create a bunch of test mailers for our specs, it makes sense to consolidate that functionality here.